+++ src/lib/components/Gift.svelte
@@ -15,7 +15,9 @@
 	import Button from './Button.svelte';
 	import Checkbox from './Checkbox.svelte';
 	import Feedback from './Feedback.svelte';
+	import Form from './Form.svelte';
 	import Options from './Options.svelte';
+	import Radio from './Radio.svelte';
 	import Slider from './Slider.svelte';
 	import TextField from './TextField.svelte';
 
@@ -57,20 +59,14 @@
 	const locale = getLocaleContext();
 </script>
 
-<form>
+<Form>
 	{#if tokens === null || tokens.length === 0}
 		<Feedback text={(l) => l.view.gift.noTokens}></Feedback>
 	{:else}
 		<fieldset>
 			<legend>{locale().view.gift.fieldset.legend}</legend>
-			<div>
-				<input bind:group={kind} type="radio" id="scholar-choice" name="scholar" value="scholar" />
-				<label for="scholar-choice">{locale().view.gift.fieldset.scholar}</label>
-			</div>
-			<div>
-				<input bind:group={kind} type="radio" id="venue-choice" name="venue" value="venue" />
-				<label for="venue-choice">{locale().view.gift.fieldset.venue}</label>
-			</div>
+			<Radio bind:group={kind} value="scholar" label={(l) => l.view.gift.fieldset.scholar} />
+			<Radio bind:group={kind} value="venue" label={(l) => l.view.gift.fieldset.venue} />
 		</fieldset>
 
 		{#if kind === 'scholar'}
@@ -131,4 +127,4 @@
 			}}
 		/>
 	{/if}
-</form>
+</Form>