+++ src/lib/components/Page.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
+	import { beforeNavigate } from '$app/navigation';
 	import { type Result } from '$lib/data/CRUD';
 	import type LocaleText from '$lib/locales/Locale';
 	import { getLocaleContext } from '$routes/Contexts';
+	import type PageHeader from '$routes/PageHeader';
 	import { getContext, type Snippet } from 'svelte';
-	import EditableText from './EditableText.svelte';
-	import Lead from './Lead.svelte';
 
 	let {
 		icon,
@@ -40,6 +40,35 @@
 	$effect(() => {
 		if (breadcrumbsContext) breadcrumbsContext.breadcrumbs = breadcrumbs;
 	});
+
+	function cleanupContext() {
+		pageHeader.icon = '';
+		pageHeader.title = '';
+		pageHeader.wobble = false;
+		pageHeader.subtitle = undefined;
+		pageHeader.details = undefined;
+		pageHeader.edit = undefined;
+	}
+
+	beforeNavigate(() => {
+		cleanupContext();
+	});
+
+	const pageHeader = getContext<PageHeader>('pageHeader');
+	$effect(() => {
+		if (pageHeader) {
+			pageHeader.icon = icon;
+			pageHeader.title = revisedTitle;
+			pageHeader.wobble = wobble;
+			pageHeader.subtitle = subtitle;
+			pageHeader.details = details;
+			pageHeader.edit = edit;
+
+			return () => {
+				cleanupContext();
+			};
+		}
+	});
 </script>
 
 <svelte:head>
@@ -47,22 +76,6 @@
 </svelte:head>
 
 <section class="page">
-	<div class="metadata">
-		<h1 class:wobble data-testid="page-header">
-			<span class="emoji">{icon}</span>
-			{#if edit}<EditableText
-					text={revisedTitle}
-					valid={edit.valid}
-					edit={edit.update}
-					strings={(l) => ({
-						placeholder: edit.placeholder(l)
-					})}
-				></EditableText>{:else}{revisedTitle}{/if}
-		</h1>
-		<div class="details">
-			{#if subtitle}<Lead>{@render subtitle()}</Lead>{/if}{@render details?.()}
-		</div>
-	</div>
 	<div class="content">
 		{@render children()}
 	</div>
@@ -77,67 +90,15 @@
 		width: 100%;
 	}
 
-	.emoji {
-		font-family: 'Noto Emoji', 'Josefin Sans', sans-serif;
-		font-size: 80%;
-	}
-
-	h1 {
-		display: flex;
-		gap: 0.5rem;
-		align-items: center;
-	}
-
-	@keyframes wobble {
-		0%,
-		100% {
-			transform: translateX(0);
-		}
-		20% {
-			transform: translateX(-5px);
-		}
-		40% {
-			transform: translateX(5px);
-		}
-		60% {
-			transform: translateX(-3px);
-		}
-		80% {
-			transform: translateX(3px);
-		}
-	}
-
-	.wobble {
-		animation: wobble 0.8s ease-in-out 0.3s 3;
-	}
-
 	.page > :global(p) {
 		margin-block-end: 0;
-	}
-
-	.metadata {
-		display: flex;
-		flex-direction: column;
-		gap: var(--spacing);
-	}
-
-	.details {
-		display: flex;
-		flex-direction: row;
-		gap: var(--spacing);
-		align-items: baseline;
-		font-size: var(--small-font-size);
-	}
-
-	.details,
-	.content {
-		padding-left: calc(var(--spacing) * 2);
-		padding-right: calc(var(--spacing) * 2);
 	}
 
 	.content {
 		display: flex;
 		flex-direction: column;
 		gap: calc(2 * var(--spacing));
+		padding-left: calc(var(--spacing) * 2);
+		padding-right: calc(var(--spacing) * 2);
 	}
 </style>