+++ src/lib/components/Tag.svelte
@@ -3,8 +3,8 @@
 <style>
 	.tag {
 		font-family: var(--font-face);
-		background: var(--salient-color);
-		color: var(--background-color);
+		background: var(--salient-color-faded);
+		color: var(--foreground-color);
 		padding: calc(var(--spacing) / 4) var(--spacing-half);
 		border-radius: 0.5em;
 		font-size: var(--extra-small-font-size);